django-pjt/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
cryptography==43.0.3
defusedxml==0.7.1
distro==1.9.0
dj-rest-auth==7.0.0
Django==4.2.16
django-allauth==0.61.1
django-cors-headers==4.6.0
djangorestframework==3.15.2
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jiter==0.7.1
oauthlib==3.2.2
openai==1.55.0
pillow==11.0.0
pycparser==2.22
pydantic==2.10.1
pydantic_core==2.27.1
PyJWT==2.9.0
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
